- 카오스던전 서브시스템 추가
적 죽을때 Subsystem으로 정화게이지 값 넘겨주는데서 크래시 발생 수정필요

+++ UPortfolio_RPG/Source/UPortfolio_RPG/Enemy/Enemy.h
@@ -85,7 +85,7 @@ public:
 	bool IsDead;
 	UPROPERTY(VisibleInstanceOnly, BlueprintReadOnly)
 	bool IsSpawn;
-	float PurificationScore = 0.f;
+	int32 PurificationScore = 0;
 
 	UEnemyAnimInstance* GetAnimInstance() { return EnemyAnim; }
 	UStatusComponent* GetStatusComponent() { return EnemyState; }
